Renamed a few packages in the sleepiq binding. (#2739)
See openhab/static-code-analysis#100
